% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{annual_growth}
\alias{annual_growth}
\title{Compute annual growth from monthly growth observations}
\usage{
annual_growth(monthly_growth)
}
\arguments{
\item{monthly_growth}{Vector of monthly growth observations where 1 means
no growth.}
}
\description{
Compute annual growth from monthly growth observations
}
