% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{b_freq}
\alias{b_freq}
\title{Empirical distribution of b}
\usage{
b_freq(trend_adjusted = FALSE)
}
\arguments{
\item{trend_adjusted}{Use trend-adjusted contraction data}
}
\description{
Empirical distribution of b
}
