% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_pc.R
\name{conditional_expectation_eqn_8}
\alias{conditional_expectation_eqn_8}
\title{Conditional expectation of equation 8}
\usage{
conditional_expectation_eqn_8(
  pc,
  s_t1,
  lambda_t,
  pi_t,
  params = ghaderi_params()
)
}
\description{
Conditional expectation of equation 8
}
