% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd.R
\name{conditional_expectation_eqn_9}
\alias{conditional_expectation_eqn_9}
\title{Conditional expectation of equation 9}
\usage{
conditional_expectation_eqn_9(
  pd,
  pc,
  s_t1,
  lambda_t,
  pi_t,
  params = ghaderi_params()
)
}
\description{
Conditional expectation of equation 9
}
