% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{E_b_trans}
\alias{E_b_trans}
\title{Expected value of transformation of b}
\usage{
E_b_trans(transform = identity, trend_adjusted = FALSE)
}
\arguments{
\item{transform}{Function taking frequencies of b as input. \code{identity}
computes the expected value of b.}

\item{trend_adjusted}{Use trend-adjusted contraction data}
}
\description{
Expected value of transformation of b
}
