% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wealth_consumption_ratio.R
\name{expectation_eqn_8}
\alias{expectation_eqn_8}
\title{\eqn{E[(1 + \exp(pc))^\theta]}}
\usage{
expectation_eqn_8(pc, lambda_t, pi_t, params = ghaderi_params())
}
\arguments{
\item{pc}{Wealth-consumption value function}

\item{lambda_t}{Intensity at time \eqn{t}}

\item{pi_t}{\eqn{P(s_t = 1 \mid \lambda_{-\infty:t})}}

\item{params}{List of parameters as returned by \link{ghaderi_params}}
}
\description{
\eqn{E[(1 + \exp(pc))^\theta]}
}
