% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_pd.R
\name{expectation_eqn_9}
\alias{expectation_eqn_9}
\title{\eqn{E[(1 + \exp(pc_{t+1}))^{\theta - 1} (1 + \exp(pd_{t+1}))]}}
\usage{
expectation_eqn_9(pc, lambda_t, pi_t, params = ghaderi_params())
}
\description{
\eqn{E[(1 + \exp(pc_{t+1}))^{\theta - 1} (1 + \exp(pd_{t+1}))]}
}
