% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_V.R
\name{expected_equity_rate}
\alias{expected_equity_rate}
\title{Eq. 9 + 10}
\usage{
expected_equity_rate(
  params = model_params(),
  conditional = FALSE,
  levered = FALSE,
  trend_adjusted = FALSE
)
}
\description{
Eq. 9 + 10
}
