% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_point_iteration.R
\name{fixed_point_iteration_pc}
\alias{fixed_point_iteration_pc}
\title{Fixed-point iteration for wealth-consumption ratio}
\usage{
fixed_point_iteration_pc(
  n_lambda = 50,
  n_pi = 20,
  params = ghaderi_params(),
  ...
)
}
\arguments{
\item{params}{List of parameters as returned by \link{ghaderi_params}}
}
\description{
Fixed-point iteration for wealth-consumption ratio
}
