% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpi.R
\name{fpi_2d}
\alias{fpi_2d}
\title{2D fixed-point iteration}
\usage{
fpi_2d(f, x, y, ..., max_iterations = 100L, abs_tolerance = 1e-04)
}
\arguments{
\item{f}{A function taking the value function and the elements of x and y
as input}

\item{x, y}{Vectors spanning the grid on which to approximate f}

\item{...}{Arguments supplied to \code{f}}

\item{max_iterations}{Maximum number of iterations}

\item{abs_tolerance}{Convergence criterion. If the average absolute change
between two consecutive iterations is less than \code{abs_tolerance} the fix-point
iteration is stopped.}
}
\description{
2D fixed-point iteration
}
