% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpi.R
\name{fpi_pc}
\alias{fpi_pc}
\title{Fixed-point iteration for wealth-consumption ratio}
\usage{
fpi_pc(params = ghaderi_params())
}
\arguments{
\item{params}{List of parameters as returned by \link{ghaderi_params}}
}
\description{
Fixed-point iteration for wealth-consumption ratio
}
