% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpi.R
\name{fpi_pd}
\alias{fpi_pd}
\title{Fixed-point iteration for price-dividend ratio}
\usage{
fpi_pd(pc_grid, params = ghaderi_params())
}
\arguments{
\item{pc_grid}{Output of \link{fpi_pc}.}

\item{params}{List of parameters as returned by \link{ghaderi_params}}
}
\description{
Fixed-point iteration for price-dividend ratio
}
