% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc.R
\name{pc_t}
\alias{pc_t}
\title{Wealth-consumption ratio}
\usage{
pc_t(pc, lambda_t, pi_t, params = ghaderi_params())
}
\arguments{
\item{pc}{Wealth-consumption value function}

\item{lambda_t}{Intensity at time \eqn{t}}

\item{pi_t}{\eqn{P(s_t = 1 \mid \lambda_{-\infty:t})}}

\item{params}{List of parameters as returned by \link{ghaderi_params}}
}
\description{
Wealth-consumption ratio
}
