% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price_dividend_ratio.R
\name{pd_t}
\alias{pd_t}
\title{Price-dividend ratio}
\usage{
pd_t(pd, lambda_t, pi_t, pc_grid, params = ghaderi_params())
}
\arguments{
\item{pd}{Price-dividend value function}

\item{lambda_t}{Intensity at time \eqn{t}}

\item{pi_t}{\eqn{P(s_t = 1 \mid \lambda_{-\infty:t})}}

\item{pc_grid}{Output of \link{fixed_point_iteration_pc}.}

\item{params}{List of parameters as returned by \link{ghaderi_params}}
}
\description{
Price-dividend ratio
}
