% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.R
\name{risk_free_rate}
\alias{risk_free_rate}
\title{Risk-free rate \eqn{r_{r,t}}}
\usage{
risk_free_rate(lambda_t, pi_t, pc_grid, params = ghaderi_params())
}
\arguments{
\item{lambda_t}{Intensity at time \eqn{t}}

\item{pi_t}{\eqn{P(s_t = 1 \mid \lambda_{-\infty:t})}}

\item{pc_grid}{Output of \link{fpi_pc}.}

\item{params}{List of parameters as returned by \link{ghaderi_params}}
}
\description{
Risk-free rate \eqn{r_{r,t}}
}
