% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulation}
\alias{simulation}
\title{Run a simulation of the model}
\usage{
simulation(s_t, pc_grid, pd_grid, params = ghaderi_params())
}
\arguments{
\item{s_t}{State variable \eqn{s_t}. Vector of zeros and ones.}

\item{params}{List of parameters as returned by \link{ghaderi_params}}
}
\description{
Run a simulation of the model
}
