% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wealth_consumption_ratio.R
\name{conditional_expectation_eqn_8}
\alias{conditional_expectation_eqn_8}
\title{Conditional expectation of equation 8}
\usage{
conditional_expectation_eqn_8(
  pc,
  s_t1,
  lambda_t,
  pi_t,
  params = ghaderi_params()
)
}
\arguments{
\item{pc}{Wealth-consumption value function}

\item{s_t1}{\eqn{s_t = 0}: normal state; \eqn{s_t = 1}: disaster state}

\item{lambda_t}{Intensity at time \eqn{t}}

\item{pi_t}{\eqn{P(s_t = 1 \mid \lambda_{-\infty:t})}}

\item{params}{List of parameters as returned by \link{ghaderi_params}}
}
\description{
Conditional expectation of equation 8
}
