% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd.R
\name{expectation_eqn_9}
\alias{expectation_eqn_9}
\title{\eqn{E[(1 + \exp(pc_{t+1}))^{\theta - 1} (1 + \exp(pd_{t+1}))]}}
\usage{
expectation_eqn_9(pd, pc, lambda_t, pi_t, params = ghaderi_params())
}
\arguments{
\item{pd}{Price-dividend value function}

\item{pc}{Wealth-consumption value function}

\item{lambda_t}{Intensity at time \eqn{t}}

\item{pi_t}{\eqn{P(s_t = 1 \mid \lambda_{-\infty:t})}}

\item{params}{List of parameters as returned by \link{ghaderi_params}}
}
\description{
\eqn{E[(1 + \exp(pc_{t+1}))^{\theta - 1} (1 + \exp(pd_{t+1}))]}
}
