% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_pi.R
\name{pi_t1}
\alias{pi_t1}
\title{Posterior belief}
\usage{
pi_t1(pi_t, lambda_t1, lambda_t, params = ghaderi_params())
}
\arguments{
\item{pi_t}{Prior belief}

\item{lambda_t1}{Current intensity}

\item{lambda_t}{Prior intensity}

\item{params}{\link{ghaderi_params}}
}
\description{
Posterior belief
}
