% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_pi.R
\name{pi_t1_t}
\alias{pi_t1_t}
\title{Probability of being in the depression state at t+1 given time-t information}
\usage{
pi_t1_t(pi_t, params = ghaderi_params())
}
\arguments{
\item{pi_t}{Probability of being in the depression state at time t}

\item{params}{\link{ghaderi_params}}
}
\description{
Probability of being in the depression state at t+1 given time-t information
}
